TDQ-5847 TDQ-5935 Backport to branch 5.1
git-svn-id: http://talendforge.org/svn/top/branches/branch-5_1@89266 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/ui/views/resources/RepositoryNodeDorpAdapterAssistant.java
@@ -338,17 +338,19 @@ public class RepositoryNodeDorpAdapterAssistant extends CommonDropAdapterAssista
 
         // get the target node and folder
         IFolder targetFolder = RepositoryNodeHelper.getIFolder(targetNode);
-        File targetFile = WorkspaceUtils.ifolderToFile(targetFolder);
+        if (targetFolder != null) {
+            File targetFile = WorkspaceUtils.ifolderToFile(targetFolder);
 
-        // move the report generate doc folder
-        FilesUtils.copyDirectory(sourceFile, targetFile);
-        FilesUtils.deleteFile(sourceFile, true);
+            // move the report generate doc folder
+            FilesUtils.copyDirectory(sourceFile, targetFile);
+            FilesUtils.deleteFile(sourceFile, true);
 
-        // update the file .report.list
-        ReportUtils.updateReportListFile(outputFolder, targetFolder);
+            // update the file .report.list
+            ReportUtils.updateReportListFile(outputFolder, targetFolder);
 
-        // refresh the dq repository tree view
-        CorePlugin.getDefault().refreshDQView(targetNode.getParent());
+            // refresh the dq repository tree view
+            CorePlugin.getDefault().refreshDQView(targetNode.getParent());
+        }
     }
 
     /**
